% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscheme_default.R
\name{colorscheme_default}
\alias{colorscheme_default}
\title{Preset blue and cyan color scheme (default)}
\usage{
colorscheme_default()
}
\value{
Nothing
}
\description{
Preset blue and cyan color scheme (default)
}
\examples{
colorscheme_default()
pioHdr("Example")
pioStr("Example", FALSE)
}
