% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscheme_orange_black.R
\name{colorscheme_orange_black}
\alias{colorscheme_orange_black}
\title{Preset yellow and black color scheme.}
\usage{
colorscheme_orange_black()
}
\value{
Nothing
}
\description{
Preset yellow and black color scheme.
}
\examples{
colorscheme_orange_black()
pioHdr("Example")
pioStr("Example", 'Test')
}
