% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorscheme_purple_ivory.R
\name{colorscheme_purple_ivory}
\alias{colorscheme_purple_ivory}
\title{Preset yellow and black color scheme.}
\usage{
colorscheme_purple_ivory()
}
\value{
Nothing
}
\description{
Preset yellow and black color scheme.
}
\examples{
colorscheme_purple_ivory()
pioHdr("Example")
pioStr("Example", 'Test')
}
