% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIO-package.R
\docType{package}
\name{PIO-package}
\alias{PIO}
\alias{PIO-package}
\title{PIO: Panel Informativity Optimizer}
\description{
PIO is an open-source software created to help researchers and biologists in designing their NGS panels.
}
\author{
\strong{Maintainer}: Vincent Alcazer \email{vincent.alczaer@chu-lyon.fr} (\href{https://orcid.org/0000-0003-1843-6286}{ORCID})

}
\keyword{internal}
