% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pioHdr.R
\name{pioHdr}
\alias{pioHdr}
\title{Print a title.}
\usage{
pioHdr(..., toPrint = NULL, prefix = "")
}
\arguments{
\item{...}{A list of parameters that will be concatenated to
create the header.}

\item{toPrint}{A named vector of arguments, that will be print in the format `n=v`
where `n` is a name and `v` a value using function \code{\link{pioStr}}.}

\item{prefix}{A prefix for each one of the strings in `toPrint`, passed to \code{\link{pioStr}}.}
}
\value{
None
}
\description{
Print a header, plus some optional parameters.
}
