% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pio.R
\name{pioStr}
\alias{pioStr}
\title{A simple string value print, with colour over string}
\usage{
pioStr(str, val, suffix = "", prefix = "")
}
\arguments{
\item{str}{First string.}

\item{val}{Second string.}
}
\value{
None
}
\description{
A simple string value print, with colour over string
}
\examples{
pioStr("My string", 12345)
pioStr("My string", TRUE)
}
