% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pio.R
\name{pioTit}
\alias{pioTit}
\title{A one-line title print to screen, surrounded by dashes}
\usage{
pioTit(title, prefix = "")
}
\arguments{
\item{title}{The title to print.}

\item{prefix}{A prefix to the string and dashes.}
}
\value{
None
}
\description{
A one-line title print to screen, surrounded by dashes
}
\examples{
pioTit('We now perform some long computation')
}
