% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pioTkn.R
\name{pioTkn}
\alias{pioTkn}
\title{Tokenizer.}
\usage{
pioTkn(..., no_color = FALSE)
}
\arguments{
\item{...}{A list of parameters to tokenize and collapse.}

\item{no_color}{If `TRUE`, avaoids tokenizing and returns
just the collapsed output string.}
}
\value{
A collapsed output strong with tokenized inputs.
}
\description{
This function tokenizes each one of
a set of ellipsis parameters, and creates a
concatenated output string with spaces. Each input
value will be tokenized as follows: if the value is
`NA` or `NULL`, a red background is reported with
white text; if the value is logical, true values will
be green and false value will be red. In all other cases
the output and the input match.
}
\examples{
# Takenize different types of variables 
cat(pioTkn(4, 5, "banana", TRUE, NULL, FALSE, NA))

# assemble without tokenizing 
cat(pioTkn(4, 5, FALSE, NA, no_color = TRUE))
}
