% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pioDisp.R
\name{pioDisp}
\alias{pioDisp}
\title{Display some content.}
\usage{
pioDisp(x)
}
\arguments{
\item{x}{Content to display}
}
\value{
Nothing.
}
\description{
This function tries to convert almost everything to `tibble`
data structure, so that we can exploit their nice layout functions.
}
\examples{
M = matrix(1:100, ncol = 10)
M[1,4] = NA
pioDisp(M)
}
