% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pio.R
\name{pioHdr}
\alias{pioHdr}
\title{Print a header, and optional parameters}
\usage{
pioHdr(header, toPrint = NULL, prefix = "", suffix = "")
}
\arguments{
\item{header}{A string}

\item{toPrint}{A named vector of arguments, that will be print in the format `n=v`
where `n` is a name and `v` a value.}

\item{prefix}{A prefix for each one of the strings in `toPrint`}

\item{suffix}{A suffix for each one of the strings in `toPrint`}
}
\value{
None
}
\description{
Print a header, and optional parameters
}
\examples{
pioHdr("My new cool function",
 toPrint = c(
  `Cutoff for significance` = 0.05,
  `Statistical test` = "Fisher",
  `Multiple Hypothesis Correction` = TRUE,
  `Plots` = NA
 ),
 prefix = '\\t-')

 # Change colours -- balck text on orange background
 options('pio.header_bg_colour' = crayon::bgYellow)
 options('pio.header_fg_colour' = crayon::black)
 pioHdr("My new cool function")
}
