% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pioStr.R
\name{pioStr}
\alias{pioStr}
\title{Print a pair of values.}
\usage{
pioStr(str, ..., suffix = "", prefix = "")
}
\arguments{
\item{str}{First string to print, the description that can
be coloured.}

\item{...}{A list of parameters that will be concatenated to
create the seconf value. These will be tokenized with \code{\link{pioTkn}}.}

\item{suffix}{Suffix to add as tail of the string.}

\item{prefix}{Suffix to add as head of the string.}
}
\value{
None
}
\description{
This function can print a pair of values
which are thought to represent a description followed by
a value. For instance `P-value = 0.05` would be print by
passing the string `"P-value ="` together with `0.05`.
The function is thought to colour the first of the two
parameters via options `pio.string_bg_colour` and
`pio.string_fg_colour`, and the second string according to
the tokenizer function \code{\link{pioTkn}}.
The function can also add a prefix and a suffix to the output,
which allows easily to indent the output message, or add
a newline at the end of it.
}
\examples{
pioStr("My string", 12345)
pioStr("My variable", TRUE, 'for some reason')
}
